[DDW-321] Separates Transactions requests into their own sub-directory
