+++ source/renderer/app/api/transactions/requests/createTransaction.js
@@ -1,11 +1,24 @@
 // @flow
-import type { AdaTransaction, AdaTransactionParams, RequestConfig } from './types';
+import type { RequestConfig } from '../../common/types';
+import type { Transaction, PaymentDistribution } from '../types';
 import { request } from '../../utils/request';
+
+export type TransactionParams = {
+  data: {
+    source: {
+      accountIndex: number,
+      walletId: string,
+    },
+    destinations: Array<PaymentDistribution>,
+    groupingPolicy: ?'OptimizeForSecurity' | 'OptimizeForSize',
+    spendingPassword: ?string
+  },
+};
 
 export const createTransaction = (
   config: RequestConfig,
-  { data }: AdaTransactionParams
-): Promise<AdaTransaction> => (
+  { data }: TransactionParams
+): Promise<Transaction> => (
   request({
     hostname: 'localhost',
     method: 'POST',
