+++ source/renderer/app/api/transactions/requests/getTransactionFee.js
@@ -1,11 +1,13 @@
 // @flow
-import type { AdaTransactionFee, AdaTxFeeParams, RequestConfig } from './types';
+import type { RequestConfig } from '../../common/types';
+import type { TransactionParams } from './createTransaction';
+import type { TransactionFee } from '../types';
 import { request } from '../../utils/request';
 
 export const getTransactionFee = (
   config: RequestConfig,
-  { data }: AdaTxFeeParams
-): Promise<AdaTransactionFee> => (
+  { data }: TransactionParams
+): Promise<TransactionFee> => (
   request({
     hostname: 'localhost',
     method: 'POST',
